refactor(Lotto): 정적 팩토리 메서드, 상수, 예외 메시지, validate()확장성 적용
