+++ src/main/java/lotto/system/ExceptionMessage.java
@@ -1,0 +1,15 @@
+package lotto.system;
+
+public enum ExceptionMessage {
+    LOTTO_LENGTH(String.format("로또 번호는 %d 가지 숫자여야 합니다", LottoNumberConstant.LENGTH.getValue()));
+
+    private final String message;
+
+    ExceptionMessage(String message) {
+        this.message = message;
+    }
+
+    public String getMessage() {
+        return "[ERROR] " + message;
+    }
+}
